src/telegram_bot/requirements.txt
fastapi
pydantic-settings
gunicorn
uvicorn
aiohttp
openai
chromadb
asyncpg
langchain
langchain-community
langchain-text-splitters
tiktoken
clean-text
unidecode
gigachat
torch
sentence-transformers
python-telegram-bot
pandas
razdel
openpyxl
pypdf
python-docx
langchain_openai